project1/requirements.txt
confluent_kafka
pandas
psycopg2-binary
kafka-python